int -> uint64_t fixes for PhysicalDisk

+++ metadata/metadata.c
@@ -1,6 +1,6 @@
 #include <metadata.h>
 
-PhysicalDisk create_physical_disk(int sector_count, int sector_size)
+PhysicalDisk create_physical_disk(uint64_t sector_count, uint64_t sector_size)
 {
         PhysicalDisk physical_disk = PHYSICAL_DISK__INIT;
         physical_disk.key = 111444890ULL;
